fixed stories
